WIP - Abstracted snapping logic into its own class
